requirements.txt
google-genai>=0.2.0
python-dotenv>=1.0.0
pillow>=9.0.0 